chore(desktop): split prompt-pipeline into prose+renderer+tree
Pure mechanical move — no behavior change. The decision tree stays
whole in prompt-pipeline.ts; only policy-data constants and pure
side-effect callees move out.

New files:
  prompt-pipeline-prose.ts (~94 lines)
    ENTER_PLAN_MODE_DESCRIPTION + PLAN_MODE_SPARSE_REMINDER constants
    with their full ADR-004 doc blocks. File header explains the seam:
    harness-supplied prose constants per ADR-004, owned here.

  prompt-pipeline-renderer.ts (~184 lines)
    emitRemoteMessageAdded, insertRendererSystemMessage,
    clearConnectingStatus, and the shared escape() helper. File header
    explains: side-effect helpers that talk to the renderer and remote
    transport; called by the decision tree but not part of its ordering
    invariants.

prompt-pipeline.ts (~452 lines):
  - Remove @file-size-exception annotation (no longer needed at this
    line count).
  - Update file-size posture section to describe the three-file cluster.
  - Add imports from the two new sibling files.
  - Re-export ENTER_PLAN_MODE_DESCRIPTION and PLAN_MODE_SPARSE_REMINDER
    from prose file so existing callers need no changes.
  - Fix duplicate /** comment block at the old handleSlash definition
    (copy-paste artifact in the original that was never flagged).

Verification: npm run typecheck clean, npm test 312/312 pass (including
prompt-pipeline.test.ts and prompt-pipeline-clear-wipe.test.ts),
make check-file-sizes OK.
